requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
cssselect==1.2.0
dnspython==2.3.0
fake-useragent==1.1.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
lxml==4.9.2
mysql-connector-python==8.0.33
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.2
parse==1.19.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyee==8.2.2
pymongo==4.3.3
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
w3lib==2.1.1
websockets==10.4
zipp==3.15.0
